+++ app/views/admin_mailer/build.html.erb
@@ -9,7 +9,7 @@
 
   links = []
   links.push([:admin_request_change_member_status.t,
-    "#{MO.http_domain}/project/change_member_status/#{@project.id}?candidate=#{@sender.id}"])
+    "#{MO.http_domain}/projects/#{@project.id}/members/#{@sender.id}/edit"])
   links.push([:email_links_show_user.t,
     "#{MO.http_domain}/users/#{@sender.id}"])
   links.push([:email_links_latest_changes.t,
