+++ app/views/layouts/sidebar/_indexes.html.erb
@@ -15,6 +15,6 @@
 <%= link_to(:NEWS.t, articles_path,
             { class: classes[:indent],
               id: "nav_articles_link" }) %>
-<%= link_to(:PROJECTS.t, project_list_projects_path,
+<%= link_to(:PROJECTS.t, projects_path,
             { class: classes[:indent],
               id: "nav_projects_link" }) %>
