+++ test/application_mailer/admin_request.html
@@ -14,7 +14,7 @@ Subject: [MO] Please do something or other
 </div>
 <div class="textile"><p>Here are some handy links:</p></div>
 <ul type=none>
-<li>Change user&#8217;s status: <a href="https://mushroomobserver.org/project/change_member_status/<%= ActiveRecord::FixtureSet.identify(:eol_project) %>?candidate=<%= ActiveRecord::FixtureSet.identify(:katrina) %>">https://mushroomobserver.org/project/change_member_status/<%= ActiveRecord::FixtureSet.identify(:eol_project) %>?candidate=<%= ActiveRecord::FixtureSet.identify(:katrina) %></a></li>
+<li>Change user&#8217;s status: <a href="https://mushroomobserver.org/projects/<%= ActiveRecord::FixtureSet.identify(:eol_project) %>/members/<%= ActiveRecord::FixtureSet.identify(:katrina) %>/edit">https://mushroomobserver.org/projects/<%= ActiveRecord::FixtureSet.identify(:eol_project) %>/members/<%= ActiveRecord::FixtureSet.identify(:katrina) %>/edit</a></li>
 <li>Show user&#8217;s profile: <a href="https://mushroomobserver.org/users/<%= ActiveRecord::FixtureSet.identify(:katrina) %>">https://mushroomobserver.org/users/<%= ActiveRecord::FixtureSet.identify(:katrina) %></a></li>
 <li>Latest changes at Mushroom Observer: <a href="https://mushroomobserver.org">https://mushroomobserver.org</a></li>
 </ul>
