+++ test/controllers/lookups_controller_test.rb
@@ -131,12 +131,12 @@ class LookupsControllerTest < FunctionalTestCase
     login
     p_id = projects(:eol_project).id
     get(:lookup_project, params: { id: p_id })
-    assert_redirected_to(controller: :project, action: :show_project, id: p_id)
+    assert_redirected_to(controller: "/projects", action: :show, id: p_id)
     get(:lookup_project, params: { id: "Bolete" })
-    assert_redirected_to(controller: :project, action: :show_project,
+    assert_redirected_to(controller: "/projects", action: :show,
                          id: projects(:bolete_project).id)
     get(:lookup_project, params: { id: "Bogus" })
-    assert_redirected_to(controller: :project, action: :index_project)
+    assert_redirected_to(controller: "/projects", action: :index)
     assert_flash_error
     get(:lookup_project, params: { id: "project" })
     # Must test against regex because passed query param borks path match
